bindKey with inputOptions for sendInput
Add a new parameter to the bindKey method to specify inputOptions.  I have a case where more than one piece of information was needed in a call to sendInput, which is easy to achieve by directly calling it in an event handler but to bind that to a keypress meant manually repeating lots of the code already in KeyboardControls class.